draw: name the Metal rail, and reach both rails through Backend
The two draw rails were `runtime::draw::encode_draw_chain` — the same path,
resolved to a different function by a `cfg`. The Vulkan rail lived in its own
file but was re-exported *flat* into `runtime::draw`, and the Metal rail was
the parent module itself with a gate on each of its nineteen items. That is why
the two arms cannot be built together: not a trait that was missing, a name that
was taken.

* `runtime/draw/metal/` is the Metal rail, the sibling of `runtime/draw/vulkan`.
  Nineteen `cfg`-gated items move into it verbatim, and `metal_icb` and
  `depth_stencil` — Metal-side already, and only reachable from those items —
  move under it as `metal::icb` and `metal::depth_stencil`. Inside the rail the
  `metal` crate is spelled `::metal`, because the module now shadows it.
* The four Metal-rail tests in `draw::tests` move to `draw::metal::tests`. They
  read that rail's private state vocabulary (`MetalStateDecline`,
  `MetalIcbInheritanceDecline`), and leaving them in a sibling would have meant
  widening those types so a test could see them.
* The Vulkan arm's `encode_icb_execute_and_writeback` — the `NoMetal` refusal
  for a rail with no `executeCommandsInBuffer:` — moves out of the shared module
  into `draw::vulkan` beside the rail it belongs to, so both rails now spell the
  entry point the same way under their own name.
* `Backend` gains `encode_draw_chain` and `encode_icb_execute_and_writeback`.
  The two rails already had byte-identical signatures over a backend-neutral
  `DrawEncodeRequest`, which is why this is a trait and not a translation layer:
  the seam was already right, it just had no name.
* `runtime::exec` calls `backend::selected().encode_draw_chain(…)` and
  `…encode_icb_execute_and_writeback(…)`. Those were the crate's only two
  product call sites into either rail, and neither names a backend now.

`force_full_store` stays a parameter of the trait method even though only Metal
reads it: the caller's reason for setting it — an abandoned chain whose partial
target must not be published — is a device fact, not a Metal one.

The Vulkan rail is still re-exported flat into `runtime::draw`; un-flattening it
is the same edit in the other direction and is kept separate so this one can be
read as the move it is.

No behavior change on either arm. Metal 1315 pass, Vulkan 2173 pass, both
serial; feature-matrix green with per-arm test counts held.
