+++ crates/reims-vgpu/src/backend/vulkan/mod.rs
@@ -22,6 +22,9 @@ pub mod engine;
 pub mod translate;
 
 use crate::backend::Backend;
+use crate::model::DeviceState;
+use crate::runtime::draw::{self, DrawEncodeRequest, EncodeStatus};
+use crate::runtime::host::{HostMemory, HostOps};
 
 /// The Vulkan rail's [`Backend`] handle.
 ///
@@ -48,6 +51,36 @@ impl Backend for VulkanBackend {
 
     fn reset(&self) {
         engine::reset_guest_state();
+    }
+
+    fn encode_draw_chain<M: HostMemory + HostOps>(
+        &self,
+        state: &mut DeviceState,
+        host: &mut M,
+        req: &mut DrawEncodeRequest,
+        writeback_guest: bool,
+        force_full_store: bool,
+    ) -> (EncodeStatus, Option<Vec<u8>>) {
+        draw::vulkan::encode_draw_chain(state, host, req, writeback_guest, force_full_store)
+    }
+
+    fn encode_icb_execute_and_writeback<M: HostMemory + HostOps>(
+        &self,
+        state: &mut DeviceState,
+        host: &mut M,
+        req: &DrawEncodeRequest,
+        icb_ref: u32,
+        range_location: u64,
+        range_length: u64,
+    ) -> EncodeStatus {
+        draw::vulkan::encode_icb_execute_and_writeback(
+            state,
+            host,
+            req,
+            icb_ref,
+            range_location,
+            range_length,
+        )
     }
 }
