+++ crates/reims-vgpu/src/runtime/compute_session.rs
@@ -819,7 +819,7 @@ fn apply_icb_compute_encoder_inheritance<M: HostMemory + HostOps>(
                     ),
                 })?;
                 // AB-resident samplers must support argument buffers.
-                let reims_vgpu = crate::runtime::draw::sampler_record(
+                let reims_vgpu = crate::runtime::draw::metal::sampler_record(
                     REIMS_VGPU_BINDING_SAMPLER_BASE + s.index,
                     &sampler.descriptor,
                     s.has_lod_clamp.then_some((s.lod_min_bits, s.lod_max_bits)),
@@ -986,7 +986,7 @@ fn apply_icb_compute_encoder_inheritance<M: HostMemory + HostOps>(
                                     ))
                                 }
                             })?;
-                    reims_vgpu_samplers.push(crate::runtime::draw::sampler_record(
+                    reims_vgpu_samplers.push(crate::runtime::draw::metal::sampler_record(
                         REIMS_VGPU_BINDING_SAMPLER_BASE + s.index,
                         &sampler.descriptor,
                         s.has_lod_clamp.then_some((s.lod_min_bits, s.lod_max_bits)),
