+++ crates/reims-vgpu/src/runtime/draw/metal/depth_stencil.rs
