+++ crates/reims-vgpu/src/runtime/draw/tests.rs
@@ -1,4 +1,9 @@
 use super::*;
+// This file tests both rails. Neither is re-exported into `super`, so each is
+// named where its arm compiles — which is also what keeps a test from asserting
+// against whichever rail happened to win a flat re-export.
+#[cfg(all(feature = "backend-metal", target_os = "macos"))]
+use super::metal::*;
 use crate::model::{DeviceId, PAGE_SHIFT_ARM64E, PAGE_SHIFT_X86};
 use crate::runtime::gva_mem::write_task_gva_arm64e;
 use crate::runtime::host::FakeHost;
@@ -1043,250 +1048,6 @@ fn encode_status_renders_its_check_beside_the_class_it_collapsed_to() {
                  class=execute recovery=metal_failed"
         );
     }
-}
-
-#[cfg(all(feature = "backend-metal", target_os = "macos"))]
-#[test]
-fn explicit_metal_sampler_and_depth_binds_return_typed_missing_entry_declines() {
-    use crate::observe::Emit;
-
-    let state = DeviceState::new(DeviceId(1), PAGE_SHIFT_ARM64E);
-    let host = FakeHost::new();
-
-    let sampler = load_sampler(&state, &host, 4, 77, 3)
-        .expect_err("a nonzero sampler ref in an empty object list must decline");
-    assert_eq!(
-        sampler,
-        MetalStateDecline::SamplerEntryMissing {
-            sampler_ref: 77,
-            index: 3,
-        }
-    );
-    assert_eq!(
-        Emit::decline("metal_draw_sampler_fallback", &sampler)
-            .field("task", 4)
-            .field("pipe", 19)
-            .field("stage", "fragment")
-            .render(),
-        format!(
-            "metal_draw_sampler_fallback reason={} \
-             sampler_ref=77 index=3 task=4 pipe=19 stage=fragment",
-            crate::observe::ladder_slug!("metal_sampler", no_list_entry)
-        )
-    );
-
-    let depth = load_depth_stencil_state(&state, &host, 4, 88)
-        .expect_err("a nonzero depth-stencil ref in an empty object list must decline");
-    assert_eq!(
-        depth,
-        MetalStateDecline::DepthStencilEntryMissing {
-            depth_stencil_ref: 88,
-        }
-    );
-    assert_eq!(
-        Emit::decline("metal_draw_depth_stencil_fallback", &depth)
-            .field("task", 4)
-            .field("pipe", 19)
-            .render(),
-        format!(
-            "metal_draw_depth_stencil_fallback reason={} \
-             depth_stencil_ref=88 task=4 pipe=19",
-            crate::observe::ladder_slug!("metal_depth_stencil", no_list_entry)
-        )
-    );
-}
-
-#[cfg(all(feature = "backend-metal", target_os = "macos"))]
-#[test]
-fn metal_state_decode_declines_delegate_the_exact_resource_decoder_reason() {
-    use crate::observe::{Decline as _, Emit};
-
-    let sampler = MetalStateDecline::SamplerDecode {
-        sampler_ref: 41,
-        index: 6,
-        reason: DecodeStatus::ErrShort("res_sampler_short"),
-    };
-    assert_eq!(sampler.slug(), "res_sampler_short");
-    assert_eq!(
-        Emit::decline("metal_draw_sampler_fallback", &sampler).render(),
-        "metal_draw_sampler_fallback reason=res_sampler_short \
-             class=short sampler_ref=41 index=6"
-    );
-
-    let depth = MetalStateDecline::DepthStencilDecode {
-        depth_stencil_ref: 52,
-        reason: DecodeStatus::ErrShort("res_depth_stencil_short"),
-    };
-    assert_eq!(depth.slug(), "res_depth_stencil_short");
-    assert_eq!(
-        Emit::decline("metal_draw_depth_stencil_fallback", &depth).render(),
-        "metal_draw_depth_stencil_fallback reason=res_depth_stencil_short \
-             class=short depth_stencil_ref=52"
-    );
-}
-
-#[cfg(all(feature = "backend-metal", target_os = "macos"))]
-#[test]
-fn every_metal_icb_inheritance_check_is_unique_namespaced_and_log_safe() {
-    use crate::observe::Decline as _;
-
-    let all = vec![
-        MetalIcbInheritanceDecline::CullModeUnsupported { value: 3 },
-        MetalIcbInheritanceDecline::FrontFacingUnsupported { value: 2 },
-        MetalIcbInheritanceDecline::BindSlotPastTable {
-            bind: PastTableBind {
-                class: BindTableClass::Buffer,
-                stage: crate::runtime::decode::render::Stage::Vertex,
-                index: MAX_BUFFER_BIND_SLOTS,
-                resource_ref: 1,
-            },
-        },
-        MetalIcbInheritanceDecline::VertexBufferMissing {
-            buffer_ref: 3,
-            index: 4,
-            offset: 5,
-        },
-        MetalIcbInheritanceDecline::FragmentBufferMissing {
-            buffer_ref: 6,
-            index: 7,
-            offset: 8,
-        },
-        MetalIcbInheritanceDecline::VertexTextureMissing {
-            texture_ref: 11,
-            index: 12,
-            detail: "no list entry".into(),
-        },
-        MetalIcbInheritanceDecline::FragmentTextureMissing {
-            texture_ref: 13,
-            index: 14,
-            detail: "guest\nread failed".into(),
-        },
-        MetalIcbInheritanceDecline::PipelineRefZero,
-        MetalIcbInheritanceDecline::PipelineMissing { pipeline_ref: 17 },
-        MetalIcbInheritanceDecline::VertexMtlbMissing { function_ref: 18 },
-        MetalIcbInheritanceDecline::FragmentMtlbMissing { function_ref: 19 },
-        MetalIcbInheritanceDecline::VertexLibraryLoad {
-            function_ref: 20,
-            detail: "Metal error".into(),
-        },
-        MetalIcbInheritanceDecline::FragmentLibraryLoad {
-            function_ref: 21,
-            detail: "Metal error".into(),
-        },
-        MetalIcbInheritanceDecline::VertexFunctionCount {
-            function_ref: 22,
-            count: 2,
-        },
-        MetalIcbInheritanceDecline::FragmentFunctionCount {
-            function_ref: 23,
-            count: 0,
-        },
-        MetalIcbInheritanceDecline::VertexFunctionGet {
-            function_ref: 24,
-            detail: "function missing".into(),
-        },
-        MetalIcbInheritanceDecline::FragmentFunctionGet {
-            function_ref: 25,
-            detail: "function missing".into(),
-        },
-        MetalIcbInheritanceDecline::VertexDescriptorMissing {
-            pipeline_ref: 26,
-            attribute_count: 3,
-        },
-        // The sibling of the entry above, and the pair is the point: an empty
-        // vertex block and a partial one are different refusals, and the
-        // builder used to report the second as a success.
-        MetalIcbInheritanceDecline::VertexAttributeUnencodable {
-            pipeline_ref: 26,
-            location: 3,
-            value: 99,
-        },
-        MetalIcbInheritanceDecline::RenderPipelineCreate {
-            pipeline_ref: 27,
-            detail: "pipeline failed".into(),
-        },
-        MetalIcbInheritanceDecline::AllocationFailed {
-            what: "inherited_texture",
-        },
-    ];
-
-    // Coverage is checked by an exhaustive `match`, not by a count. A literal
-    // here is a second spelling of the variant list: when the enum lost five
-    // checks this assertion still said 26 and failed for that reason alone,
-    // reporting a shrunk vocabulary as a fixture gap. `variant_name` has no `_`
-    // arm, so a new check stops this test compiling until the fixture carries
-    // one — which is what "the fixture must cover every check" was asking for.
-    fn variant_name(decline: &MetalIcbInheritanceDecline) -> &'static str {
-        use MetalIcbInheritanceDecline as D;
-        match decline {
-            D::CullModeUnsupported { .. } => "CullModeUnsupported",
-            D::FrontFacingUnsupported { .. } => "FrontFacingUnsupported",
-            D::BindSlotPastTable { .. } => "BindSlotPastTable",
-            D::VertexBufferMissing { .. } => "VertexBufferMissing",
-            D::FragmentBufferMissing { .. } => "FragmentBufferMissing",
-            D::VertexTextureMissing { .. } => "VertexTextureMissing",
-            D::FragmentTextureMissing { .. } => "FragmentTextureMissing",
-            D::PipelineRefZero => "PipelineRefZero",
-            D::PipelineMissing { .. } => "PipelineMissing",
-            D::VertexMtlbMissing { .. } => "VertexMtlbMissing",
-            D::FragmentMtlbMissing { .. } => "FragmentMtlbMissing",
-            D::VertexLibraryLoad { .. } => "VertexLibraryLoad",
-            D::FragmentLibraryLoad { .. } => "FragmentLibraryLoad",
-            D::VertexFunctionCount { .. } => "VertexFunctionCount",
-            D::FragmentFunctionCount { .. } => "FragmentFunctionCount",
-            D::VertexFunctionGet { .. } => "VertexFunctionGet",
-            D::FragmentFunctionGet { .. } => "FragmentFunctionGet",
-            D::VertexDescriptorMissing { .. } => "VertexDescriptorMissing",
-            D::VertexAttributeUnencodable { .. } => "VertexAttributeUnencodable",
-            D::RenderPipelineCreate { .. } => "RenderPipelineCreate",
-            D::AllocationFailed { .. } => "AllocationFailed",
-        }
-    }
-    let mut covered = all.iter().map(variant_name).collect::<Vec<_>>();
-    covered.sort_unstable();
-    covered.dedup();
-    assert_eq!(
-        covered.len(),
-        all.len(),
-        "the fixture names one check twice: {covered:?}"
-    );
-
-    let mut slugs = all.iter().map(Decline::slug).collect::<Vec<_>>();
-    for decline in &all {
-        assert!(decline.slug().starts_with("metal_icb_inherit_"));
-        for (key, value) in decline.fields() {
-            assert!(!key.is_empty());
-            assert!(
-                !value.chars().any(char::is_whitespace),
-                "{} rendered a non-token field {key}={value:?}",
-                decline.slug()
-            );
-        }
-    }
-    slugs.sort_unstable();
-    slugs.dedup();
-    assert_eq!(slugs.len(), all.len(), "two ICB checks share one reason");
-}
-
-#[cfg(all(feature = "backend-metal", target_os = "macos"))]
-#[test]
-fn metal_icb_inheritance_line_keeps_pipeline_and_sanitized_driver_detail() {
-    use crate::observe::Emit;
-
-    let decline = MetalIcbInheritanceDecline::RenderPipelineCreate {
-        pipeline_ref: 71,
-        detail: "Error Domain=MTLLibrary Code=3".into(),
-    };
-    assert_eq!(
-        Emit::decline("metal_icb_inheritance", &decline)
-            .field("task", 2)
-            .field("pipe", 71)
-            .field("icb", 99)
-            .render(),
-        "metal_icb_inheritance reason=metal_icb_inherit_render_pipeline_create \
-             pipeline_ref=71 detail=Error_Domain=MTLLibrary_Code=3 \
-             task=2 pipe=71 icb=99"
-    );
 }
 
 /// A vertex reflection that trips the shader-pull coverage gate: writes
