+++ crates/reims-vgpu/src/runtime/icb/tests.rs
@@ -26,9 +26,9 @@ use crate::runtime::decode::resource::{
     OBJECT_TYPE_FUNCTION, PIPELINE_TAG_KERNEL_FUNC, TYPE7_FIRST_TLVS, TYPE7_OBJECT_COMPUTE_PIPELINE,
 };
 #[cfg(all(feature = "backend-metal", target_os = "macos"))]
-use crate::runtime::draw::{
-    encode_icb_execute_and_writeback, BufferBind, ColorRtRequest, DrawEncodeRequest, EncodeStatus,
-};
+use crate::runtime::draw::metal::encode_icb_execute_and_writeback;
+#[cfg(all(feature = "backend-metal", target_os = "macos"))]
+use crate::runtime::draw::{BufferBind, ColorRtRequest, DrawEncodeRequest, EncodeStatus};
 use crate::runtime::gva_mem;
 use crate::runtime::host::FakeHost;
 /// Readback, the draw encoder and the fixture directory. Every Metal-arm ICB
